Set default settings and display the settings of a specific image

+++ Library/Classes/ImageClass.cs
@@ -8,15 +8,16 @@
 	public class ImageSettings
 	{
 		public int width { get; set; }
-		public int s_brightness { get; set; }
-		public int s_contrast { get; set; }
-		public int s_saturation { get; set; }
+		public float s_brightness { get; set; }
+		public float s_contrast { get; set; }
+		public float s_saturation { get; set; }
 		public string s_grayscale { get; set; }
 		public bool s_sepia { get; set; }
 		public string sflip { get; set; }
 		public string watermark { get; set; }
 		public string title { get; set; }
 	}
+
 
 	public class ImageConfig
 	{
